Moar algorithm
Progress:

There is now an arbitrary hypothesis-making algorithm, allowing for the solving of anything.

Class ranges: each constraint can have one (1) range of values. For a sudoku, normally it's 1-9 , or in the case of a clue it's n-n where n is the clue itself.

Technical:

Strong typedefs. There are some limitations (names are case sensitive and a maximum of 13 characters long) but otherwise they work fine. And the limitations can be lifted easily.

The strong typedefs are used for the ClassValue and ClassID classes, since those two must not be mixed accidentally.

The Ops namespace is now much smaller since I do most operations elsewhere.